pkg/encryption/text_encryption.go
package encryption

import (
	"crypto/aes"
	"crypto/cipher"
	"crypto/rand"
	"encoding/base64"
	"errors"
	"io"

	"github.com/gogatekeeper/gatekeeper/pkg/apperrors"
)

// encryptDataBlock encrypts the plaintext string with the key
func EncryptDataBlock(plaintext, key []byte) ([]byte, error) {
	block, err := aes.NewCipher(key)

	if err != nil {
		return []byte{}, err
	}

	gcm, err := cipher.NewGCM(block)

	if err != nil {
		return []byte{}, err
	}

	nonce := make([]byte, gcm.NonceSize())

	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
		return nil, err
	}

	return gcm.Seal(nonce, nonce, plaintext, nil), nil
}

// decryptDataBlock decrypts some cipher text
func DecryptDataBlock(cipherText, key []byte) ([]byte, error) {
	block, err := aes.NewCipher(key)

	if err != nil {
		return []byte{}, err
	}

	gcm, err := cipher.NewGCM(block)

	if err != nil {
		return []byte{}, err
	}

	nonceSize := gcm.NonceSize()

	if len(cipherText) < nonceSize {
		return nil, errors.New("failed to decrypt the ciphertext, the text is too short")
	}

	nonce, input := cipherText[:nonceSize], cipherText[nonceSize:]

	return gcm.Open(nil, nonce, input, nil)
}

// encodeText encodes the session state information into a value for a cookie to consume
func EncodeText(plaintext string, key string) (string, error) {
	cipherText, err := EncryptDataBlock([]byte(plaintext), []byte(key))

	if err != nil {
		return "", err
	}

	return base64.RawStdEncoding.EncodeToString(cipherText), nil
}

// decodeText decodes the session state cookie value
func DecodeText(state, key string) (string, error) {
	cipherText, err := base64.RawStdEncoding.DecodeString(state)

	if err != nil {
		return "", err
	}
	// step: decrypt the cookie back in the expiration|token
	encoded, err := DecryptDataBlock(cipherText, []byte(key))

	if err != nil {
		return "", apperrors.ErrInvalidSession
	}

	return string(encoded), nil
}
